Replace SimpleConnectionPool implementation with the improved connection pool.
