+++ components/client/src/main/java/com/hotels/styx/client/connectionpool/ConnectionPools.java
@@ -15,13 +15,13 @@
  */
 package com.hotels.styx.client.connectionpool;
 
-import com.hotels.styx.api.extension.Origin;
 import com.hotels.styx.api.MetricRegistry;
+import com.hotels.styx.api.extension.Origin;
 import com.hotels.styx.api.extension.service.BackendService;
 import com.hotels.styx.client.netty.connectionpool.NettyConnectionFactory;
 
-import static com.hotels.styx.client.HttpRequestOperationFactory.Builder.httpRequestOperationFactoryBuilder;
 import static com.hotels.styx.api.extension.service.ConnectionPoolSettings.defaultConnectionPoolSettings;
+import static com.hotels.styx.client.HttpRequestOperationFactory.Builder.httpRequestOperationFactoryBuilder;
 
 /**
  * Helper routines for building connection pools with default settings.
