+++ components/client/src/test/unit/java/com/hotels/styx/client/connectionpool/SimpleConnectionPoolTest.java
@@ -15,641 +15,716 @@
  */
 package com.hotels.styx.client.connectionpool;
 
-import com.hotels.styx.api.Id;
-import com.hotels.styx.api.exceptions.TransportException;
+import com.hotels.styx.api.LiveHttpRequest;
+import com.hotels.styx.api.LiveHttpResponse;
+import com.hotels.styx.api.exceptions.OriginUnreachableException;
 import com.hotels.styx.api.extension.Origin;
 import com.hotels.styx.api.extension.service.ConnectionPoolSettings;
 import com.hotels.styx.client.Connection;
 import com.hotels.styx.client.ConnectionSettings;
-import com.hotels.styx.client.connectionpool.stubs.StubConnectionFactory;
-import com.hotels.styx.client.connectionpool.stubs.StubConnectionFactory.StubConnection;
+import org.reactivestreams.Publisher;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+import reactor.test.StepVerifier;
 import rx.Observable;
-import rx.Subscriber;
-import rx.observers.TestSubscriber;
+import rx.subjects.PublishSubject;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.atomic.AtomicReference;
+import java.time.Duration;
+import java.util.Random;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import java.util.concurrent.atomic.AtomicLong;
 
-import static com.hotels.styx.api.Id.id;
 import static com.hotels.styx.api.extension.Origin.newOriginBuilder;
-import static com.hotels.styx.client.connectionpool.ConnectionPoolStatsCounter.NULL_CONNECTION_POOL_STATS;
-import static com.hotels.styx.support.api.BlockingObservables.getFirst;
-import static com.hotels.styx.support.matchers.RegExMatcher.matchesRegex;
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
-import static org.hamcrest.CoreMatchers.instanceOf;
-import static org.hamcrest.CoreMatchers.not;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.Is.is;
-import static org.hamcrest.core.IsNull.notNullValue;
-import static org.hamcrest.core.IsNull.nullValue;
+import static com.hotels.styx.api.extension.service.ConnectionPoolSettings.defaultConnectionPoolSettings;
+import static java.lang.String.format;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static rx.Observable.error;
-import static rx.Observable.just;
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertFalse;
+import static org.testng.Assert.assertTrue;
 
 public class SimpleConnectionPoolTest {
-    private static final Id ORIGIN = id("foo");
-
-    static final ActionOnOpen FAIL_TO_CONNECT = () -> {
-        throw new TransportException("");
-    };
-
-    Connection.Factory connectionFactory = new StubConnectionFactory();
-    SimpleConnectionPool connectionPool;
-    ConnectionPoolSettings config;
+    private final Origin origin = newOriginBuilder("localhost", 9090).build();
+    private final ConnectionSettings settings = new ConnectionSettings(1000);
+    private Connection.Factory connectionFactory;
+    private Connection connection1;
+    private Connection connection2;
+    private Connection connection3;
+    private Connection connection4;
 
     @BeforeMethod
-    public void makeHostConnectPool() {
-        config = connectionPoolConfig().build();
-        connectionPool = originConnectionPool(ORIGIN, "localhost", 9090, connectionFactory, config);
+    public void setUp() {
+        connectionFactory = mock(Connection.Factory.class);
+
+        connection1 = mock(Connection.class);
+        when(connection1.isConnected()).thenReturn(true);
+
+        connection2 = mock(Connection.class);
+        when(connection2.isConnected()).thenReturn(true);
+
+        connection3 = mock(Connection.class);
+        when(connection3.isConnected()).thenReturn(true);
+
+        connection4 = mock(Connection.class);
+        when(connection4.isConnected()).thenReturn(true);
     }
 
     @Test
-    public void closingPoolWillCloseBorrowedConnection() {
-        Connection connection = borrowConnectionSynchronously();
-        Connection.Listener listener = mock(Connection.Listener.class);
-        connection.addConnectionListener(listener);
+    public void borrowsConnection() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class))).thenReturn(Observable.just(connection1));
 
-        connectionPool.close();
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, defaultConnectionPoolSettings(), connectionFactory);
 
-        verify(listener).connectionClosed(connection);
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
+
+        assertEquals(pool.stats().connectionAttempts(), 1);
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+        assertEquals(pool.stats().pendingConnectionCount(), 0);
     }
 
     @Test
-    public void closingPoolWillCloseAvailableConnection() {
-        Connection connection = borrowConnectionSynchronously();
-        Connection.Listener listener = mock(Connection.Listener.class);
-        connection.addConnectionListener(listener);
+    public void borrowsReturnedConnection() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.just(connection2));
 
-        connectionPool.returnConnection(connection);
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, defaultConnectionPoolSettings(), connectionFactory);
 
-        connectionPool.close();
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
 
-        verify(listener).connectionClosed(connection);
+        assertEquals(pool.stats().connectionAttempts(), 1);
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+
+        pool.returnConnection(connection1);
+
+        assertEquals(pool.stats().busyConnectionCount(), 0);
+        assertEquals(pool.stats().availableConnectionCount(), 1);
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
+
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+    }
+
+
+    @Test
+    public void emitsConnectionWhenOnceEstablished() throws ExecutionException, InterruptedException {
+        PublishSubject<Connection> subject = PublishSubject.create();
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(subject);
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, defaultConnectionPoolSettings(), connectionFactory);
+
+        // The pool is empty, so the `borrowConnection` triggers a connection establishment.
+        Publisher<Connection> future = pool.borrowConnection2();
+
+        StepVerifier.create(future)
+                .then(() -> {
+                    assertEquals(pool.stats().pendingConnectionCount(), 1);
+                    assertEquals(pool.stats().busyConnectionCount(), 0);
+                    assertEquals(pool.stats().availableConnectionCount(), 0);
+                })
+                .expectNextCount(0)
+                .then(() -> {
+                    subject.onNext(connection1);
+                    subject.onCompleted();
+                })
+                .expectNext(connection1)
+                .verifyComplete();
+
+        assertEquals(pool.stats().connectionAttempts(), 1);
+        assertEquals(pool.stats().pendingConnectionCount(), 0);
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+    }
+
+
+    @Test
+    public void purgesTerminatedConnections() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.just(connection2));
+
+        // Create a new connection
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, defaultConnectionPoolSettings(), connectionFactory);
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .then(() -> pool.returnConnection(connection1))
+                .verifyComplete();
+
+        when(connection1.isConnected()).thenReturn(false);
+        pool.connectionClosed(connection1);
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection2)
+                .verifyComplete();
+
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 0);
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+
+        assertEquals(pool.stats().terminatedConnections(), 1);
+        assertEquals(pool.stats().closedConnections(), 0);
     }
 
     @Test
-    public void closingPoolWithTerminateSubscribersWithAnError() throws InterruptedException {
-        StubConnectionFactory connectionFactory = new StubConnectionFactory();
-        connectionPool = new SimpleConnectionPool(
-                origin(),
-                connectionPoolConfig()
-                        .maxConnectionsPerHost(1)
-                        .maxPendingConnectionsPerHost(1)
-                        .connectTimeout(1, MILLISECONDS)
-                        .build(),
-                connectionFactory);
+    public void returnsConnectionToWaitingSubscribers() throws ExecutionException, InterruptedException {
+        PublishSubject<Connection> subject = PublishSubject.create();
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(subject);
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, defaultConnectionPoolSettings(), connectionFactory);
 
-        borrowConnectionSynchronously();
+        // The pool is empty, so the `borrowConnection` triggers a connection establishment.
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
 
-        Observable<Connection> observable = connectionPool.borrowConnection();
+        // Another borrow comes in, and triggers a connection establishment.
+        // The subject keeps the connection perpetually in "3-way handshake"
+        Publisher<Connection> publisher = pool.borrowConnection2();
+        CompletableFuture<Connection> future = Mono.from(publisher).toFuture();
 
-        AtomicReference<Connection> connRef = new AtomicReference<>();
-        AtomicReference<Throwable> errorRef = new AtomicReference<>();
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 1);
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
 
-        observable.subscribe(new Subscriber<Connection>() {
-            @Override
-            public void onCompleted() {
+        // The returned connection is now given to the waiting subscriber:
+        pool.returnConnection(connection1);
 
-            }
+        assertTrue(future.isDone());
+        assertEquals(future.get(), connection1);
 
-            @Override
-            public void onError(Throwable e) {
-                errorRef.set(e);
-            }
-
-            @Override
-            public void onNext(Connection connection) {
-                connRef.set(connection);
-            }
-        });
-
-        connectionPool.close();
-
-        assertThat(connRef.get(), is(nullValue()));
-        assertThat(errorRef.get(), is(notNullValue()));
-
-        Throwable error = errorRef.get();
-
-        assertThat(error.getMessage(), is("Connection pool closed"));
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 0);
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
     }
 
     @Test
-    public void canRoundTripBorrowedConnections() {
-        Connection borrowedConnection = borrowConnectionSynchronously();
+    public void returnConnectionChecksIfConnectionIsClosed() {
+        PublishSubject<Connection> subject = PublishSubject.create();
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(subject);
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, defaultConnectionPoolSettings(), connectionFactory);
 
-        assertThat(borrowedConnection, is(notNullValue()));
-        assertThat(connectionPool.stats().busyConnectionCount(), is(1));
+        // The pool is empty, so the `borrowConnection` triggers a connection establishment.
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
 
-        boolean connectionClosed = connectionPool.returnConnection(borrowedConnection);
-        assertThat(connectionClosed, is(false));
-        assertThat(connectionPool.stats().busyConnectionCount(), is(0));
-        assertThat(connectionPool.stats().availableConnectionCount(), is(1));
+        // Another borrow comes in, and triggers a connection establishment.
+        // The subject keeps the connection perpetually in "3-way handshake"
+        CompletableFuture<Connection> future = Mono.from(pool.borrowConnection2()).toFuture();
+        assertFalse(future.isDone());
+
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 1);
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+
+        // The returned connection is closed, and then returned:
+        when(connection1.isConnected()).thenReturn(false);
+        pool.returnConnection(connection1);
+
+        // The closed connection is discarded, and the subscriber still waits for the 3-way handshake
+        assertFalse(future.isDone());
+
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 1);
+        assertEquals(pool.stats().busyConnectionCount(), 0);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+        assertEquals(pool.stats().terminatedConnections(), 0);
+        assertEquals(pool.stats().closedConnections(), 0);
+
+        // Finally the connectionClosed callback fires:
+        pool.connectionClosed(connection1);
+
+        assertEquals(pool.stats().terminatedConnections(), 1);
+        assertEquals(pool.stats().closedConnections(), 0);
     }
 
+
     @Test
-    public void returnsAConnectionFromPoolIfPresent() {
-        Connection borrowedConnection = borrowConnectionSynchronously();
+    public void unsubscribingRemovesTheWaitingSubscriber() throws ExecutionException, InterruptedException {
+        PublishSubject<Connection> subject1 = PublishSubject.create();
+        PublishSubject<Connection> subject2 = PublishSubject.create();
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(subject1)
+                .thenReturn(subject2);
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, defaultConnectionPoolSettings(), connectionFactory);
 
-        connectionPool.returnConnection(borrowedConnection);
+        // The subscriber is pending because the connection is still 3-way handshaking
+        CompletableFuture<Connection> future1 = Mono.from(pool.borrowConnection2()).toFuture();
 
-        Connection secondConnection = borrowConnectionSynchronously();
-        assertThat(secondConnection, is(not(nullValue())));
-        assertThat(connectionPool.stats().busyConnectionCount(), is(1));
-        assertThat(connectionPool.stats().availableConnectionCount(), is(0));
+        // Borrow another one:
+        CompletableFuture<Connection> future2 = Mono.from(pool.borrowConnection2()).toFuture();
+
+        assertEquals(pool.stats().pendingConnectionCount(), 2);
+
+        // `Future1` is ahead of `future2` in the waiting subscribers.
+        // Cancel the future1 and verify that `future2` gets it spot:
+        future1.cancel(true);
+
+        assertEquals(pool.stats().pendingConnectionCount(), 1);
+
+        subject1.onNext(connection1);
+        subject1.onCompleted();
+
+        assertEquals(pool.stats().pendingConnectionCount(), 0);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+
+        assertTrue(future1.isCompletedExceptionally());
+        assertEquals(future2.get(), connection1);
     }
 
+
+    // 2. The connection pool limits
     @Test
-    public void updatesStatsOnConnectionError() {
-        Connection.Factory connectionFactory = connectionFactory(FAIL_TO_CONNECT);
-        this.connectionPool = connectionPool(connectionPoolConfig(), connectionFactory);
+    public void limitsPendingConnectionsDueToConnectionEstablishment() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(PublishSubject.create())
+                .thenReturn(PublishSubject.create())
+                .thenReturn(PublishSubject.create());
 
-        TestSubscriber<Connection> subscriber = new TestSubscriber<>();
-        this.connectionPool.borrowConnection().subscribe(subscriber);
-
-        assertThat(subscriber.getOnErrorEvents().get(0), instanceOf(TransportException.class));
-        assertThat(connectionPool.stats().pendingConnectionCount(), is(0));
-        assertThat(connectionPool.stats().busyConnectionCount(), is(0));
-    }
-
-    @Test
-    public void opensANewConnectionWhenNoAvailableFromPool() {
-        Connection firstConnection = borrowConnectionSynchronously();
-        assertThat(connectionPool.stats().busyConnectionCount(), is(1));
-        Connection secondConnection = borrowConnectionSynchronously();
-
-        assertThat(firstConnection, is(notNullValue()));
-        assertThat(secondConnection, is(notNullValue()));
-
-        assertThat(connectionPool.stats().busyConnectionCount(), is(2));
-    }
-
-    @Test
-    public void disregardsReturnConnectionIfTheConnectionIsAlreadyClosed() throws IOException {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(2)
-                .connectTimeout(1, MILLISECONDS));
-
-        Connection connection = borrowConnectionSynchronously();
-        connectionPool.closeConnection(connection);
-
-        connectionPool.returnConnection(connection);
-
-        assertThat(this.connectionPool.stats().availableConnectionCount(), is(0));
-        assertThat(this.connectionPool.stats().busyConnectionCount(), is(0));
-    }
-
-    @Test
-    public void disregardsCloseConnectionIfTheConnectionIsAlreadyReturned() throws IOException {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(2)
-                .connectTimeout(1, MILLISECONDS));
-
-        Connection connection = borrowConnectionSynchronously();
-        connectionPool.returnConnection(connection);
-
-        connectionPool.closeConnection(connection);
-
-        assertThat(this.connectionPool.stats().availableConnectionCount(), is(1));
-    }
-
-    @Test
-    public void discardsTheReturnedConnectionIfNotConnected() throws IOException {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(2)
-                .connectTimeout(1, MILLISECONDS));
-
-        Connection oldConnection = borrowConnectionSynchronously();
-        connectionPool.closeConnection(oldConnection);
-
-        connectionPool.returnConnection(oldConnection);
-
-        assertThat(this.connectionPool.stats().availableConnectionCount(), is(0));
-        assertThat(this.connectionPool.stats().busyConnectionCount(), is(0));
-    }
-
-    @Test
-    public void willNotCloseTheReturnedConnectionIfHealthy() throws IOException {
-        Connection connection = mock(Connection.class);
-        when(connection.isConnected()).thenReturn(true);
-
-        connectionPool = originConnectionPool(ORIGIN, "localhost", 9090, connectionFactory, config);
-        connectionPool.returnConnection(connection);
-        verify(connection, never()).close();
-    }
-
-    @Test
-    public void willNotOpenANewConnectionIfPendingOpenConnectionsIsOverTheMaximumConfigured() {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(1)
-                .maxPendingConnectionsPerHost(0));
-
-        TestSubscriber<Connection> subscriber = new TestSubscriber<>();
-        this.connectionPool.borrowConnection().subscribe(subscriber);
-        this.connectionPool.borrowConnection().subscribe(subscriber);
-
-        Exception exception = (Exception) subscriber.getOnErrorEvents().get(0);
-        assertThat(exception, is(instanceOf(MaxPendingConnectionsExceededException.class)));
-        assertThat(exception.getMessage(), matchesRegex("" +
-                "Maximum allowed pending connections exceeded for origin=foo:h1:localhost:9090. " +
-                "pendingConnectionsCount=[0-9] is greater than maxPendingConnectionsPerHost=[0-9].*"));
-    }
-
-    @Test
-    public void willPutConnectionRequestInWaitingListIfBusyConnectionCountOverTheMaximumConfigured() {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(1)
-                .maxPendingConnectionsPerHost(1)
-                .connectTimeout(1, MILLISECONDS));
-
-        borrowConnectionSynchronously();
-        this.connectionPool.borrowConnection().subscribe(new TestSubscriber<>());
-
-        assertThat(connectionPool.stats().availableConnectionCount(), is(0));
-        assertThat(connectionPool.stats().pendingConnectionCount(), is(1));
-    }
-
-    @Test
-    public void doesNotBorrowDeadConnectionIfConnectionDiesAfterReturning() throws IOException {
-        Connection firstConnection = borrowConnectionSynchronously();
-        connectionPool.returnConnection(firstConnection);
-        firstConnection.close();
-
-        Connection secondConnection = borrowConnectionSynchronously();
-        assertThat(secondConnection.isConnected(), is(true));
-    }
-
-    @Test
-    public void subscriberWillNotBeOfferedConnectionOnceItsBeenNotifiedOfAnError() {
-        Connection.Factory connectionFactory = connectionFactory(FAIL_TO_CONNECT);
-        this.connectionPool = connectionPool(connectionPoolConfig(), connectionFactory);
-
-        TestSubscriber subscriber = new TestSubscriber();
-        connectionPool.borrowConnection().subscribe(subscriber);
-
-        assertThat(subscriber.getOnErrorEvents().get(0), is(instanceOf(Throwable.class)));
-        assertThat(this.connectionPool.stats().pendingConnectionCount(), is(0));
-    }
-
-    @Test
-    public void willNotHandoutClosedConnection() throws IOException {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(3)
-                .connectTimeout(1, MILLISECONDS));
-
-        borrowReturnAndCloseConnection();
-        borrowReturnAndCloseConnection();
-
-        Connection connection = borrowConnectionSynchronously();
-        assertThat(connection.isConnected(), is(true));
-    }
-
-    @Test
-    public void removesAvailableConnectionsFromPoolWhenTheyTerminate() throws Exception {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(2)
-                .connectTimeout(1, MILLISECONDS));
-
-        List<Connection> connections = populatePool(this.connectionPool, 2);
-        closeAll(connections);
-
-        assertThat(this.connectionPool.stats().availableConnectionCount(), is(0));
-    }
-
-    @Test
-    public void borrowingAllTheConnectionsWillExhaustThePool() {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(1)
-                .maxPendingConnectionsPerHost(1)
-                .connectTimeout(1, MILLISECONDS));
-
-        // max out the pool
-        borrowConnectionSynchronously();
-        // max out the pending connections
-        this.connectionPool.borrowConnection().subscribe(new TestSubscriber<>());
-
-        assertThat(connectionPool.isExhausted(), is(true));
-    }
-
-    @Test
-    public void timesOutLongestWaitingSubscribers() {
-        TestSubscriber<Connection> subscriber = new TestSubscriber<>();
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                        .maxConnectionsPerHost(1)
-                        .pendingConnectionTimeout(100, MILLISECONDS)
-        );
-
-        borrowConnectionSynchronously();
-        this.connectionPool.borrowConnection().subscribe(subscriber);
-
-        subscriber.awaitTerminalEvent();
-        assertThat(subscriber.getOnErrorEvents().get(0), instanceOf(MaxPendingConnectionTimeoutException.class));
-    }
-
-    @Test
-    public void doesNotApplyTimeoutAfterEmittingResponse() {
-        TestSubscriber<Connection> subscriber = new TestSubscriber<>();
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                        .maxConnectionsPerHost(1)
-                        .pendingConnectionTimeout(100, MILLISECONDS)
-        );
-
-        this.connectionPool.borrowConnection().subscribe(subscriber);
-
-        subscriber.awaitTerminalEvent(500, MILLISECONDS);
-        assertThat(connectionBorrowed(subscriber), instanceOf(Connection.class));
-        subscriber.assertNoErrors();
-    }
-
-    @Test
-    public void acceptsTtfbValueOfZero() throws IOException {
-        Connection connection = mock(Connection.class);
-        when(connection.isConnected()).thenReturn(true);
-        when(connection.getTimeToFirstByteMillis()).thenReturn(0L);
-
-        connectionPool = originConnectionPool(ORIGIN, "localhost", 9090, connectionFactory, config);
-        connectionPool.returnConnection(connection);
-        verify(connection, never()).close();
-    }
-
-    @Test
-    public void shouldCreateAPoolWithoutStats() {
-        Origin anyOrigin = newOriginBuilder("localhost", 9090)
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
+                .maxPendingConnectionsPerHost(2)
                 .build();
-        ConnectionPool pool = new SimpleConnectionPool(
-                anyOrigin, new ConnectionPoolSettings.Builder().build(), new StubConnectionFactory(), false);
 
-        assertThat(pool.stats(), is(NULL_CONNECTION_POOL_STATS));
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        CompletableFuture<Connection> pending1 = Mono.from(pool.borrowConnection2()).toFuture();
+        CompletableFuture<Connection> pending2 = Mono.from(pool.borrowConnection2()).toFuture();
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectError(MaxPendingConnectionsExceededException.class)
+                .verify();
+
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 2);
+
     }
 
+    // 2. The connection pool limits
     @Test
-    public void willNotRecordNegativeLatencyValues() {
-        Origin anyOrigin = newOriginBuilder("localhost", 9090)
-                .build();
-        Connection.Factory factory = (origin, connectionPoolConfiguration) -> {
-            Connection stubConnection = new StubConnection(origin) {
-                @Override
-                public long getTimeToFirstByteMillis() {
-                    return -1L;
-                }
-            };
-            return just(stubConnection);
-        };
+    public void limitsPendingConnectionsDueToPoolSaturation() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.just(connection2))
+                .thenReturn(Observable.just(connection3))
+                .thenReturn(Observable.just(connection4));
 
-        ConnectionPool pool = new SimpleConnectionPool(anyOrigin, new ConnectionPoolSettings.Builder().build(), factory, true);
-
-        Connection connection = borrowConnectionSynchronously(pool);
-        pool.returnConnection(connection);
-
-        assertThat(pool.stats().timeToFirstByteMs(), is(0L));
-    }
-
-    @Test
-    public void shouldTimeoutAndLogAfterConfiguredPendingTimeout() {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(1)
-                .maxPendingConnectionsPerHost(1)
-                .pendingConnectionTimeout(100, MILLISECONDS)
-                .connectTimeout(1, MILLISECONDS));
-
-        borrowConnectionSynchronously();
-        TestSubscriber<Connection> subscriber = new TestSubscriber<>();
-        this.connectionPool.borrowConnection().subscribe(subscriber);
-        subscriber.awaitTerminalEvent();
-
-        assertThat(connectionPool.stats().availableConnectionCount(), is(0));
-        assertThat(connectionPool.stats().pendingConnectionCount(), is(0));
-
-        assertThat(errorMessage(subscriber), is("Maximum wait time exceeded for origin=foo:h1:localhost:9090. pendingConnectionTimeoutMillis=100"));
-    }
-
-    @Test
-    public void willTakeConnectionFromWaitingListIfBusyConnectionCountDropsBelowTheMaximumConfigured() {
-        this.connectionPool = connectionPool(connectionPoolConfig()
-                .maxConnectionsPerHost(1)
-                .maxPendingConnectionsPerHost(1)
-                .connectTimeout(1, MILLISECONDS));
-
-        Connection connection = borrowConnectionSynchronously();
-        TestSubscriber<Connection> subscriber = new TestSubscriber<>();
-        this.connectionPool.borrowConnection().subscribe(subscriber);
-
-        assertThat(connectionPool.stats().availableConnectionCount(), is(0));
-        assertThat(connectionPool.stats().pendingConnectionCount(), is(1));
-
-        connectionPool.returnConnection(connection);
-
-        assertThat(connectionPool.stats().pendingConnectionCount(), is(0));
-        assertThat(connectionBorrowed(subscriber), is(connection));
-        assertThat("busy connections", connectionPool.stats().busyConnectionCount(), is(1));
-    }
-
-    @Test
-    public void returnsConnectionsAfterUse() {
-        Connection.Factory factory = mockConnectionFactory();
-        SimpleConnectionPool simpleConnectionPool = sizeOnePool(factory);
-        AtomicReference<Connection> firstConnection = new AtomicReference<>();
-
-        Observable<String> result = simpleConnectionPool.withConnection(connection -> {
-            firstConnection.set(connection);
-            return just("foo");
-        });
-
-        assertThat(result.toBlocking().single(), is("foo"));
-        assertThat(simpleConnectionPool.stats().busyConnectionCount(), is(0));
-        assertThat(simpleConnectionPool.stats().availableConnectionCount(), is(1));
-        assertThat(simpleConnectionPool.borrowConnection().toBlocking().single(), is(firstConnection.get()));
-    }
-
-    @Test
-    public void closesConnectionUponError() {
-        Connection.Factory factory = mockConnectionFactory();
-        SimpleConnectionPool simpleConnectionPool = sizeOnePool(factory);
-        AtomicReference<Connection> firstConnection = new AtomicReference<>();
-
-        Observable<String> result = simpleConnectionPool.withConnection(connection -> {
-            firstConnection.set(connection);
-            return error(new RuntimeException());
-        });
-
-        TestSubscriber<String> subscriber = new TestSubscriber<>();
-        result.subscribe(subscriber);
-        subscriber.assertError(RuntimeException.class);
-
-        assertThat(simpleConnectionPool.stats().busyConnectionCount(), is(0));
-        assertThat(simpleConnectionPool.stats().availableConnectionCount(), is(0));
-        assertThat(simpleConnectionPool.borrowConnection().toBlocking().single(), is(not(firstConnection.get())));
-    }
-
-    @Test
-    public void countsConnectionEstablishmentAttempts() throws Exception {
-        Connection borrowedConnection = borrowConnectionSynchronously();
-        assertThat(borrowedConnection, is(notNullValue()));
-        assertThat(connectionPool.stats().connectionAttempts(), is(1));
-
-        boolean connectionClosed = connectionPool.returnConnection(borrowedConnection);
-        assertThat(connectionClosed, is(false));
-
-        borrowConnectionSynchronously();
-        assertThat(connectionPool.stats().connectionAttempts(), is(1));
-    }
-
-    @Test
-    public void countsConnectionEstablishmentFailures() throws Exception {
-        Connection.Factory connectionFactory = connectionFactory(FAIL_TO_CONNECT);
-        this.connectionPool = connectionPool(connectionPoolConfig(), connectionFactory);
-
-        TestSubscriber<Connection> subscriber = new TestSubscriber<>();
-        this.connectionPool.borrowConnection().subscribe(subscriber);
-
-        assertThat(connectionPool.stats().connectionFailures(), is(1));
-    }
-
-    @Test
-    public void countsClosedConnections() throws Exception {
-        Connection connection = borrowConnectionSynchronously();
-        connectionPool.closeConnection(connection);
-
-        assertThat(this.connectionPool.stats().closedConnections(), is(1));
-        assertThat(this.connectionPool.stats().terminatedConnections(), is(1));
-    }
-
-    @Test
-    public void countsTerminatedConnections() throws Exception {
-        this.connectionPool = connectionPool(connectionPoolConfig()
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
                 .maxConnectionsPerHost(2)
-                .connectTimeout(1, MILLISECONDS));
+                .maxPendingConnectionsPerHost(2)
+                .build();
 
-        Connection connection = borrowConnectionSynchronously();
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
 
-        connection.close();
+        // Saturate active connections:
+        CompletableFuture<Connection> pending1 = Mono.from(pool.borrowConnection2()).toFuture();
+        CompletableFuture<Connection> pending2 = Mono.from(pool.borrowConnection2()).toFuture();
+        assertTrue(pending1.isDone());
+        assertTrue(pending2.isDone());
 
-        assertThat(this.connectionPool.stats().terminatedConnections(), is(1));
-        assertThat(this.connectionPool.stats().closedConnections(), is(0));
-        assertThat("busy connections ", this.connectionPool.stats().busyConnectionCount(), is(0));
+        // These are pending
+        CompletableFuture<Connection> pending3 = Mono.from(pool.borrowConnection2()).toFuture();
+        CompletableFuture<Connection> pending4 = Mono.from(pool.borrowConnection2()).toFuture();
 
-        connectionPool.returnConnection(connection);
+        assertFalse(pending3.isDone());
+        assertFalse(pending4.isDone());
 
-        assertThat(this.connectionPool.stats().terminatedConnections(), is(1));
-        assertThat(this.connectionPool.stats().closedConnections(), is(0));
-        assertThat("busy connections ", this.connectionPool.stats().busyConnectionCount(), is(0));
+        // This throws an error:
+        StepVerifier.create(pool.borrowConnection2())
+                .expectError(MaxPendingConnectionsExceededException.class)
+                .verify();
+
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 2);
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+        assertEquals(pool.stats().terminatedConnections(), 0);
+        assertEquals(pool.stats().closedConnections(), 0);
+    }
+
+
+    // 2. The connection pool limits
+    @Test
+    public void returnConnectionDecrementsConnectionCount() throws ExecutionException, InterruptedException, TimeoutException {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.just(connection2))
+                .thenReturn(Observable.just(connection3))
+                .thenReturn(Observable.just(connection4));
+
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
+                .maxConnectionsPerHost(2)
+                .maxPendingConnectionsPerHost(2)
+                .build();
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        // Saturate active connections:
+        CompletableFuture<Connection> active1 = Mono.from(pool.borrowConnection2()).toFuture();
+        CompletableFuture<Connection> active2 = Mono.from(pool.borrowConnection2()).toFuture();
+
+        assertTrue(active1.isDone());
+        assertTrue(active2.isDone());
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+
+        // Create two pending connections
+        CompletableFuture<Connection> pending1 = Mono.from(pool.borrowConnection2()).toFuture();
+        CompletableFuture<Connection> pending2 = Mono.from(pool.borrowConnection2()).toFuture();
+
+        assertFalse(pending1.isDone());
+        assertFalse(pending2.isDone());
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 2);
+
+        // Return active connections
+        pool.returnConnection(active1.get());
+        pool.returnConnection(active2.get());
+
+        // And ensure pending connections complete
+        assertEquals(pending1.get(), connection1);
+        assertEquals(pending2.get(), connection2);
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 0);
+
+        // Returning connections ...
+        pool.returnConnection(connection1);
+        pool.returnConnection(connection2);
+
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().busyConnectionCount(), 0);
+        assertEquals(pool.stats().pendingConnectionCount(), 0);
+
+        // .. allows more connections to be borrowed:
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection2)
+                .verifyComplete();
+
+        assertEquals(pool.stats().connectionAttempts(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 0);
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+        assertEquals(pool.stats().terminatedConnections(), 0);
+        assertEquals(pool.stats().closedConnections(), 0);
     }
 
     @Test
-    private SimpleConnectionPool sizeOnePool(Connection.Factory factory) {
-        return new SimpleConnectionPool.Factory()
-                .connectionPoolSettings(new ConnectionPoolSettings(1, 1, 1000, 1000, 1L))
-                .connectionFactory(factory)
-                .create(origin());
-    }
+    public void closesConnections() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.just(connection2))
+                .thenReturn(Observable.just(connection3))
+                .thenReturn(Observable.just(connection4));
 
-    private Connection.Factory mockConnectionFactory() {
-        Connection.Factory factory = mock(Connection.Factory.class);
-        when(factory.createConnection(any(Origin.class), any(ConnectionSettings.class))).thenAnswer(invocation -> {
-            Connection mock = mock(Connection.class);
-            when(mock.isConnected()).thenReturn(true);
-            return just(mock);
-        });
-        return factory;
-    }
-
-    private Connection connectionBorrowed(TestSubscriber<Connection> subscriber) {
-        return subscriber.getOnNextEvents().get(0);
-    }
-
-    private SimpleConnectionPool connectionPool(ConnectionPoolSettings.Builder configuration) {
-        return originConnectionPool(ORIGIN, "localhost", 9090, connectionFactory, configuration.build());
-    }
-
-    private List<Connection> populatePool(ConnectionPool hostConnectionPool, int numConnections) {
-        List<Connection> connections = new ArrayList<>(numConnections);
-
-        for (int i = 0; i < numConnections; i++) {
-            connections.add(borrowConnectionSynchronously());
-        }
-
-        for (Connection connection : connections) {
-            hostConnectionPool.returnConnection(connection);
-        }
-
-        return connections;
-    }
-
-    private void borrowReturnAndCloseConnection() throws IOException {
-        Connection firstConnection = borrowConnectionSynchronously();
-        connectionPool.returnConnection(firstConnection);
-        firstConnection.close();
-    }
-
-    private Connection borrowConnectionSynchronously() {
-        Connection connection = getFirst(connectionPool.borrowConnection());
-        assertThat(connection, is(notNullValue()));
-        return connection;
-    }
-
-    private Connection borrowConnectionSynchronously(ConnectionPool connectionPool) {
-        Connection connection = getFirst(connectionPool.borrowConnection());
-        assertThat(connection, is(notNullValue()));
-        return connection;
-    }
-
-    private static String errorMessage(TestSubscriber<Connection> subscriber) {
-        return subscriber.getOnErrorEvents().get(0).getMessage();
-    }
-
-    private static SimpleConnectionPool connectionPool(ConnectionPoolSettings.Builder configuration, Connection.Factory connectionFactory) {
-        return originConnectionPool(ORIGIN, "localhost", 9090, connectionFactory, configuration.build());
-    }
-
-    private static SimpleConnectionPool originConnectionPool(Id applicationId, String host, int port, Connection.Factory factory,
-                                                             ConnectionPoolSettings settings) {
-        return new SimpleConnectionPool(
-                newOriginBuilder(host, port)
-                        .applicationId(applicationId)
-                        .id("h1")
-                        .build(), settings, factory);
-    }
-
-    private static Origin origin() {
-        return newOriginBuilder("localhost", 9090)
-                .applicationId(ORIGIN)
-                .id("h1")
-                .build();
-    }
-
-    private static void closeAll(Iterable<Connection> connections) throws IOException {
-        for (Connection connection : connections) {
-            connection.close();
-        }
-    }
-
-    private static Connection.Factory connectionFactory(ActionOnOpen actionOnOpen) {
-        return (origin, connectionPoolConfiguration) ->
-                Observable.create(subscriber -> {
-                    try {
-                        actionOnOpen.execute();
-                        subscriber.onNext(new StubConnection(origin));
-                        subscriber.onCompleted();
-                    } catch (TransportException e) {
-                        subscriber.onError(e);
-                    }
-                });
-    }
-
-    private static ConnectionPoolSettings.Builder connectionPoolConfig() {
-        return new ConnectionPoolSettings.Builder()
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
                 .maxConnectionsPerHost(2)
-                .connectTimeout(200, MILLISECONDS);
+                .maxPendingConnectionsPerHost(2)
+                .build();
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
+
+        pool.closeConnection(connection1);
+
+        verify(connection1).close();
+        assertEquals(pool.stats().terminatedConnections(), 0);
+        assertEquals(pool.stats().closedConnections(), 1);
     }
 
-    interface ActionOnOpen {
-        void execute() throws TransportException;
+    @Test
+    public void closeConnectionDecrementsBorrowedCount() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.just(connection2))
+                .thenReturn(Observable.just(connection3))
+                .thenReturn(Observable.just(connection4));
+
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
+                .maxConnectionsPerHost(2)
+                .maxPendingConnectionsPerHost(2)
+                .build();
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        // Saturate pool:
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection2)
+                .verifyComplete();
+
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().terminatedConnections(), 0);
+        assertEquals(pool.stats().closedConnections(), 0);
+
+        // Close a connection
+        pool.closeConnection(connection1);
+
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().closedConnections(), 1);
+
+        // A new connection has been established ...
+        // proving that the borrowed count was decremented
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection3)
+                .verifyComplete();
+
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().terminatedConnections(), 0);
+        assertEquals(pool.stats().closedConnections(), 1);
+    }
+
+
+    @Test
+    public void closeConnectionTriggersConnectionEstablishment() throws ExecutionException, InterruptedException {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.just(connection2))
+                .thenReturn(Observable.just(connection3))
+                .thenReturn(Observable.just(connection4));
+
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
+                .maxConnectionsPerHost(2)
+                .maxPendingConnectionsPerHost(2)
+                .build();
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        // Saturate pool:
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection2)
+                .verifyComplete();
+
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+
+        // Create a pending connection
+        CompletableFuture<Connection> pending1 = Mono.from(pool.borrowConnection2()).toFuture();
+        assertFalse(pending1.isDone());
+
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+
+        // Close a previously borrowed connection:
+        pool.closeConnection(connection1);
+
+        // A new connection is established and handed over to pending subscriber:
+        assertEquals(pending1.get(), connection3);
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().pendingConnectionCount(), 0);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+        assertEquals(pool.stats().closedConnections(), 1);
+    }
+
+
+    @Test
+    public void idleActiveConnectionMakesRoomForOthers() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.just(connection2))
+                .thenReturn(Observable.just(connection3))
+                .thenReturn(Observable.just(connection4));
+
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
+                .maxConnectionsPerHost(2)
+                .maxPendingConnectionsPerHost(2)
+                .build();
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        // Create a new connection
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .verifyComplete();
+
+        assertEquals(pool.stats().busyConnectionCount(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+
+        // Put it back to pool
+        pool.returnConnection(connection1);
+
+        assertEquals(pool.stats().busyConnectionCount(), 0);
+        assertEquals(pool.stats().availableConnectionCount(), 1);
+
+        // Connection terminates while it is in active queue
+        // Note: `connectionClosed` is a notification, NOT a formal API for returning a connection.
+        // therefore `availableConnections` metric stays at 1.
+        when(connection1.isConnected()).thenReturn(false);
+        pool.connectionClosed(connection1);
+
+        assertEquals(pool.stats().busyConnectionCount(), 0);
+        assertEquals(pool.stats().availableConnectionCount(), 1);
+        assertEquals(pool.stats().closedConnections(), 0);
+        assertEquals(pool.stats().terminatedConnections(), 1);
+
+        // Connection closure doesn't affect subsequent borrows:
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection2)
+                .verifyComplete();
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection3)
+                .verifyComplete();
+
+        assertEquals(pool.stats().busyConnectionCount(), 2);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+        assertEquals(pool.stats().closedConnections(), 0);
+        assertEquals(pool.stats().terminatedConnections(), 1);
+    }
+
+    @Test
+    public void connectionEstablishmentFailureRetryThreeTimesAtBorrow() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())))
+                .thenReturn(Observable.just(connection4));
+
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
+                .maxConnectionsPerHost(2)
+                .maxPendingConnectionsPerHost(2)
+                .build();
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection4)
+                .verifyComplete();
+    }
+
+    @Test
+    public void connectionEstablishmentFailureRetryThreeTimesAtConnectionClosure() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())))
+                .thenReturn(Observable.just(connection4));
+
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
+                .maxConnectionsPerHost(2)
+                .maxPendingConnectionsPerHost(2)
+                .build();
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .then(() -> {
+                    assertEquals(pool.stats().availableConnectionCount(), 0);
+                    assertEquals(pool.stats().closedConnections(), 0);
+                    assertEquals(pool.stats().pendingConnectionCount(), 0);
+                    assertEquals(pool.stats().busyConnectionCount(), 1);
+                })
+                .then(() -> pool.closeConnection(connection1))
+                .then(() -> {
+                    assertEquals(pool.stats().availableConnectionCount(), 1);
+                    assertEquals(pool.stats().closedConnections(), 1);
+                    assertEquals(pool.stats().pendingConnectionCount(), 0);
+                    assertEquals(pool.stats().busyConnectionCount(), 0);
+                })
+                .verifyComplete();
+    }
+
+    @Test
+    public void connectionEstablishmentFailureRetryThreeTimesOnlyAtBorrow() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())));
+
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
+                .maxConnectionsPerHost(2)
+                .maxPendingConnectionsPerHost(2)
+                .build();
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        Mono.from(pool.borrowConnection2()).subscribe();
+
+        assertEquals(pool.stats().pendingConnectionCount(), 1);
+        assertEquals(pool.stats().connectionFailures(), 1);
+        assertEquals(pool.stats().availableConnectionCount(), 0);
+        assertEquals(pool.stats().closedConnections(), 0);
+        assertEquals(pool.stats().terminatedConnections(), 0);
+    }
+
+    @Test
+    public void connectionEstablishmentFailureRetryThreeTimesOnlyAtConnectionClosure() {
+        when(connectionFactory.createConnection(any(Origin.class), any(ConnectionSettings.class)))
+                .thenReturn(Observable.just(connection1))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())))
+                .thenReturn(Observable.error(new OriginUnreachableException(origin, new RuntimeException())));
+
+        ConnectionPoolSettings poolSettings = new ConnectionPoolSettings.Builder()
+                .maxConnectionsPerHost(2)
+                .maxPendingConnectionsPerHost(2)
+                .build();
+
+        SimpleConnectionPool pool = new SimpleConnectionPool(origin, poolSettings, connectionFactory);
+
+        StepVerifier.create(pool.borrowConnection2())
+                .expectNext(connection1)
+                .then(() -> {
+                    assertEquals(pool.stats().availableConnectionCount(), 0);
+                    assertEquals(pool.stats().closedConnections(), 0);
+                    assertEquals(pool.stats().pendingConnectionCount(), 0);
+                    assertEquals(pool.stats().busyConnectionCount(), 1);
+                })
+                .then(() -> {
+                    pool.closeConnection(connection1);
+                    pool.connectionClosed(connection1);
+                })
+                .then(() -> {
+                    assertEquals(pool.stats().connectionFailures(), 1);
+                    assertEquals(pool.stats().pendingConnectionCount(), 0);
+                    assertEquals(pool.stats().availableConnectionCount(), 0);
+                    assertEquals(pool.stats().closedConnections(), 1);
+                    assertEquals(pool.stats().terminatedConnections(), 1);
+                })
+                .verifyComplete();
     }
 }
